+++ Database/DatabaseUpdater.cs
@@ -29,7 +29,7 @@ public sealed class DatabaseUpdater
         _versionPath = versionPath;
         _logger = logger;
         _http = new HttpClient { Timeout = TimeSpan.FromSeconds(20) };
-        _http.DefaultRequestHeaders.UserAgent.Add(new ProductInfoHeaderValue("TrueAutoHDR", "1.2.6"));
+        _http.DefaultRequestHeaders.UserAgent.Add(new ProductInfoHeaderValue("TrueAutoHDR", "1.3.0"));
     }
 
     public string CurrentVersion